one Saga element for every data
