--- src/features/getMovieData/movieDetailsSaga.js
@@ -1,18 +1,0 @@
-import { call, delay, put, takeEvery, takeLatest } from "redux-saga/effects";
-import { getMoviesDetails } from "./getData";
-import { setLoading, setMovieDetails } from "./MovieSlice/movieSlice";
-
-export function* fetchMovieDetailsWorker({payload: movieId}) {
-  try {
-    yield delay(100);
-    const movieDetails = yield call(getMoviesDetails, movieId);
-    yield put(setMovieDetails(movieDetails));
-  } catch (error) {
-    yield call(alert("coś poszło nie tak! Spróbuj później :)"));
-  }
-}
-
-
-export function* movieDetailsSaga() {
-  yield takeLatest(setLoading.type, fetchMovieDetailsWorker);
-}
