+++ src/features/getMovieData/movieSaga.js
@@ -1,8 +1,9 @@
 import { call, delay, put, takeLatest, takeLeading } from "redux-saga/effects";
-import { getPopularMovies, getPopularPeople, getGenres } from "./getData";
+import { getPopularMovies, getPopularPeople, getGenres, getMoviesDetails } from "./getData";
 import { fetchPopularMovies, setMovieList } from "./MovieSlice/movieSlice";
 import { fetchGenres, setGenreList } from "./MovieSlice/movieSlice";
 import { fetchPopularPeople, setPeopleList } from "./PeopleSlice/peopleSlice";
+import { setLoading, setMovieDetails } from "./MovieSlice/movieSlice";
 
 export function* fetchPopularMoviesWorker() {
   try {
@@ -32,6 +33,15 @@ export function* fetchPopularPeopleWorker() {
     yield call(alert("coś poszło nie tak! Spróbuj później :)"));
   }
 }
+export function* fetchMovieDetailsWorker({payload: movieId}) {
+  try {
+    yield delay(100);
+    const movieDetails = yield call(getMoviesDetails, movieId);
+    yield put(setMovieDetails(movieDetails));
+  } catch (error) {
+    yield call(alert("coś poszło nie tak! Spróbuj później :)"));
+  }
+}
 export function* movieSaga() {
   yield takeLatest(fetchPopularMovies.type, fetchPopularMoviesWorker);
   yield takeLatest(fetchGenres.type, fetchGenresWorker);
@@ -40,3 +50,15 @@ export function* movieSaga() {
 export function* peopleSaga() {
   yield takeLatest(fetchPopularPeople.type, fetchPopularPeopleWorker);
 }
+export function* movieDetailsSaga() {
+  yield takeLatest(setLoading.type, fetchMovieDetailsWorker);
+}
+
+
+
+
+
+
+
+
+
