+++ src/features/getMovieData/store.js
@@ -3,7 +3,7 @@ import createSagaMiddleware from "@redux-saga/core";
 import movieListReducer from "./MovieSlice/movieSlice";
 import peopleListReducer from "./PeopleSlice/peopleSlice";
 import genreListReducer from "./MovieSlice/movieSlice";
-//import movieDetailsReducer from "../getMovieData/MovieDetails/movieDetailsSlice";
+import movieDetailsReducer from "../getMovieData/MovieSlice/movieSlice";
 import rootSaga from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
@@ -13,6 +13,7 @@ const store = configureStore({
     movieList: movieListReducer,
     peopleList: peopleListReducer,
     genreList: genreListReducer,
+    movieDetails: movieDetailsReducer,
   },
   middleware: [sagaMiddleware],
 });
